Added a status column to pickupRequests and updated readme

+++ auth/pickupRequestsRouter.js
@@ -1,3 +1,6 @@
+//Had to do too much logic that could break with small changes to the db.
+//If I were to do this over, I'd seperate pickup requests into two or more tables
+
 const express = require('express');
 const tokenAuth = require('./authUtils/tokenAuthentication.js');
 const db = require('../data/dbModel.js');
@@ -21,12 +24,12 @@ router.post('/', (req, res) => {
 					res.status(201).json({ message: 'Pickup request created!' });
 				})
 				.catch((err) => {
-					res.status(500).json({ message: 'Server error', error: err });
+					res.status(500).json({ message: 'Server error', error: err.message });
 				});
 		} else {
-			res
-				.status(400)
-				.json({ message: 'Provide a foodType, amount and preferredPickupTime.' });
+			res.status(400).json({
+				message: 'Provide a foodType, amount and preferredPickupTime.',
+			});
 		}
 	} else {
 		res.status(401).json({
@@ -41,7 +44,7 @@ router.get('/', (req, res) => {
 			res.json(returned);
 		})
 		.catch((err) => {
-			res.status(500).json({ message: 'Server error', error: err });
+			res.status(500).json({ message: 'Server error', error: err.message });
 		});
 });
 
@@ -55,7 +58,7 @@ router.put('/:id', (req, res) => {
 						db.updatePickupRequest(req.params.id, {
 							foodType: req.body.foodType,
 							amount: req.body.amount,
-							preferredPickupTime: req.body.preferredPickupTime
+							preferredPickupTime: req.body.preferredPickupTime,
 						})
 							.then((returned) => {
 								res.status(200).json({
@@ -65,7 +68,7 @@ router.put('/:id', (req, res) => {
 							.catch((err) => {
 								res.status(500).json({
 									message: 'Server error',
-									error: err,
+									error: err.message,
 								});
 							});
 					} else {
@@ -76,37 +79,87 @@ router.put('/:id', (req, res) => {
 					}
 				})
 				.catch((err) => {
-					res.status(500).json({ message: 'Server error', error: err });
-				});
-		} else {
-			res
-				.status(400)
-				.json({ message: 'Provide a foodType, amount and preferredPickupTime.' });
-		}
-	} else if (req.decodedToken.accountType === 'volunteer') {
-		if (req.body.assign === true) {
-			db.updatePickupRequest(req.params.id, {
-				volunteerAccountID: req.decodedToken.userID,
-			})
-				.then((returned) => {
-					res.status(200).json({ message: 'Pickup request updated!' });
-				})
-				.catch((err) => {
-					res.status(500).json({ message: 'Server error', error: err });
-				});
-		} else if (req.body.assign === false) { //Add a check to see if they're even the assigned volunteer first
-			db.updatePickupRequest(req.params.id, {
-				volunteerAccountID: null,
-			})
-				.then((returned) => {
-					res.status(200).json({ message: 'Pickup request updated!' });
-				})
-				.catch((err) => {
-					res.status(500).json({ message: 'Server error', error: err });
+					res.status(500).json({ message: 'Server error', error: err.message });
 				});
 		} else {
 			res.status(400).json({
-				message: 'Provide an object with "assign" set to true or false.',
+				message: 'Provide a foodType, amount and preferredPickupTime.',
+			});
+		}
+	} else if (req.decodedToken.accountType === 'volunteer') {
+		if (req.body.status === 'Assigned') {
+			db.findPickupById(req.params.id)
+				.then((found) => {
+					//Add check to see if it's not already assigned or complete
+					if (found.status === 'Pending') {
+						db.updatePickupRequest(req.params.id, {
+							volunteerAccountID: req.decodedToken.userID,
+							status: 'Assigned',
+						}).then((returned) => {
+							res.status(200).json({ message: 'Pickup request updated!' });
+						});
+					} else {
+						res.json({
+							message:
+								'You do not have authorization to update this pickup request. The pickup request is already assigned.',
+						});
+					}
+				})
+				.catch((err) => {
+					res.status(500).json({
+						message: 'Server error',
+						error: err.message,
+					});
+				});
+		} else if (
+			req.body.status === 'Pending' ||
+			req.body.status === 'Complete'
+		) {
+			db.findPickupById(req.params.id)
+				.then((found) => {
+					//Add a check to see if they're even the assigned volunteer first & that the pickup request isn't complete yet
+					if (
+						found.volunteerAccountID === req.decodedToken.userID &&
+						found.status != 'Complete'
+					) {
+						switch (req.body.status) {
+							//If attempting to cancel the pickup request, it sets the volunteer value to null and changes status to pending
+							case 'Pending':
+								db.updatePickupRequest(req.params.id, {
+									volunteerAccountID: null,
+									status: 'Pending',
+								}).then((returned) => {
+									res.status(200).json({
+										message: 'Pickup request updated!',
+									});
+								});
+								break;
+							//If attempting to complete the pickup request, it sets it to complete
+							case 'Complete':
+								db.updatePickupRequest(req.params.id, {
+									status: 'Complete',
+								}).then((returned) => {
+									res.status(200).json({
+										message: 'Pickup request updated!',
+									});
+								});
+						}
+					} else {
+						res.json({
+							message:
+								'You do not have authorization to update this pickup request. The pickup request may not be assigned to you or is already complete.',
+						});
+					}
+				})
+				.catch((err) => {
+					res.status(500).json({
+						message: 'Server error',
+						error: err.message,
+					});
+				});
+		} else {
+			res.status(400).json({
+				message: 'Provide an object with "status" set to Complete or Pending.',
 			});
 		}
 	}
@@ -117,7 +170,10 @@ router.delete('/:id', (req, res) => {
 		//Check if this biz made the pickup request first before letting them delete it
 		db.findPickupById(req.params.id)
 			.then((found) => {
-				if (found.businessAccountID === req.decodedToken.userID) {
+				if (
+					found.businessAccountID === req.decodedToken.userID &&
+					found.status !== 'Complete'
+				) {
 					db.deletePickupRequest(req.params.id)
 						.then((returned) => {
 							res.status(200).json({ message: 'Pickup request deleted!' });
@@ -125,41 +181,49 @@ router.delete('/:id', (req, res) => {
 						.catch((err) => {
 							res.status(500).json({
 								message: 'Failed to delete pickup request',
-								error: err,
-							});
-						});
-				}
-			})
-			.catch((err) => {
-				res.status(500).json({ message: 'Server error', error: err });
-			});
-	} else if (req.decodedToken.accountType === 'volunteer') {
-		//Only allow a volunteer to delete (aka complete) a pickup if they're assigned to it
-		db.findPickupById(req.params.id)
-			.then((found) => {
-				if (found.volunteerAccountID === req.decodedToken.userID) {
-					console.log('equalk');
-					db.deletePickupRequest(req.params.id)
-						.then((returned) => {
-							res.status(200).json({ message: 'Pickup request deleted!' });
-						})
-						.catch((err) => {
-							res.status(500).json({
-								message: 'Failed to delete pickup request',
-								error: err,
+								error: err.message,
 							});
 						});
 				} else {
-					res.status(401).json({
-						message:
-							'You do not have authorization to delete this pickup request.',
-					});
+					res.status(401).json({ message: 'You do not have authorization to delete this pickup request.' });
 				}
 			})
 			.catch((err) => {
-				res.status(500).json({ message: 'Server error', error: err });
+				res.status(500).json({ message: 'Server error', error: err.message });
 			});
+	} else {
+		res.status(401).json({
+			message: 'You do not have authorization to delete this pickup request.',
+		});
 	}
+	// else if (req.decodedToken.accountType === 'volunteer') {
+	// 	//Only allow a volunteer to delete (aka complete) a pickup if they're assigned to it
+	// 	db.findPickupById(req.params.id)
+	// 		.then((found) => {
+	// 			if (found.volunteerAccountID === req.decodedToken.userID) {
+	// 				console.log('equalk');
+	// 				db.deletePickupRequest(req.params.id)
+	// 					.then((returned) => {
+	// 						res.status(200).json({ message: 'Pickup request deleted!' });
+	// 					})
+	// 					.catch((err) => {
+	// 						res.status(500).json({
+	// 							message: 'Failed to delete pickup request',
+	// 							error: err.message,
+	// 						});
+	// 					});
+	// 			}
+	// 			else {
+	// 				res.status(401).json({
+	// 					message:
+	// 						'You do not have authorization to delete this pickup request.',
+	// 				});
+	// 			}
+	// 		})
+	// 		.catch((err) => {
+	// 			res.status(500).json({ message: 'Server error', error: err });
+	// 		});
+	// }
 });
 
 module.exports = router;
